+++ frontend/src/components/actuators/FanCard.vue
